+++ client-impl/simple-client/src/main/java/uk/gov/gchq/palisade/client/SimpleClient.java
@@ -18,27 +18,15 @@ package uk.gov.gchq.palisade.client;
 
 import uk.gov.gchq.palisade.Context;
 import uk.gov.gchq.palisade.UserId;
-import uk.gov.gchq.palisade.audit.service.AuditService;
-import uk.gov.gchq.palisade.audit.service.impl.LoggerAuditService;
-import uk.gov.gchq.palisade.cache.service.CacheService;
-import uk.gov.gchq.palisade.cache.service.impl.HashMapCacheService;
 import uk.gov.gchq.palisade.data.serialise.NullSerialiser;
 import uk.gov.gchq.palisade.data.serialise.Serialiser;
 import uk.gov.gchq.palisade.data.service.DataService;
 import uk.gov.gchq.palisade.data.service.request.ReadRequest;
 import uk.gov.gchq.palisade.data.service.request.ReadResponse;
-import uk.gov.gchq.palisade.policy.service.PolicyService;
-import uk.gov.gchq.palisade.policy.service.impl.HashMapPolicyService;
 import uk.gov.gchq.palisade.resource.Resource;
-import uk.gov.gchq.palisade.resource.service.HashMapResourceService;
-import uk.gov.gchq.palisade.resource.service.ResourceService;
-import uk.gov.gchq.palisade.service.PalisadeService;
-import uk.gov.gchq.palisade.service.impl.SimplePalisadeService;
 import uk.gov.gchq.palisade.service.request.ConnectionDetail;
 import uk.gov.gchq.palisade.service.request.DataRequestResponse;
 import uk.gov.gchq.palisade.service.request.RegisterDataRequest;
-import uk.gov.gchq.palisade.user.service.UserService;
-import uk.gov.gchq.palisade.user.service.impl.HashMapUserService;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -47,29 +35,20 @@ import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
 import java.util.stream.Stream;
 
-public class SimpleClient<T> implements SimpleServices {
+public class SimpleClient<T> {
     private final Serialiser<T> serialiser;
-    private final ResourceService resourceService;
-    private final AuditService auditService;
-    private final PolicyService policyService;
-    private final UserService userService;
-    private final CacheService cacheService;
-    private final PalisadeService palisadeService;
+    private final ServicesFactory services;
 
-    public SimpleClient() {
+    public SimpleClient(final ServicesFactory services) {
+        Objects.requireNonNull(services);
+        this.services = services;
         this.serialiser = createSerialiser();
-        this.resourceService = createResourceService();
-        this.auditService = createAuditService();
-        this.policyService = createPolicyService();
-        this.userService = createUserService();
-        this.cacheService = createCacheService();
-        this.palisadeService = createPalisadeService();
     }
 
     public Stream<T> read(final String filename, final String resourceType, final String userId, final String justification) {
-        Objects.requireNonNull(getPalisadeService());
+        Objects.requireNonNull(services.getPalisadeService());
         final RegisterDataRequest dataRequest = new RegisterDataRequest().resourceId(filename).userId(new UserId().id(userId)).context(new Context().justification(justification));
-        final DataRequestResponse dataRequestResponse = getPalisadeService().registerDataRequest(dataRequest).join();
+        final DataRequestResponse dataRequestResponse = services.getPalisadeService().registerDataRequest(dataRequest).join();
         final List<CompletableFuture<Stream<T>>> futureResults = new ArrayList<>(dataRequestResponse.getResources().size());
 
         for (final Entry<Resource, ConnectionDetail> entry : dataRequestResponse.getResources().entrySet()) {
@@ -90,72 +69,11 @@ public class SimpleClient<T> implements SimpleServices {
         return futureResults.stream().flatMap(CompletableFuture::join);
     }
 
-    protected PalisadeService createPalisadeService() {
-        return new SimplePalisadeService(
-                getResourceService(),
-                getAuditService(),
-                getPolicyService(),
-                getUserService(),
-                getCacheService()
-        );
-    }
-
-    protected CacheService createCacheService() {
-        return new HashMapCacheService();
-    }
-
-    protected ResourceService createResourceService() {
-        return new HashMapResourceService();
-    }
-
-    protected AuditService createAuditService() {
-        return new LoggerAuditService();
-    }
-
-    protected PolicyService createPolicyService() {
-        return new HashMapPolicyService();
-    }
-
-
-    protected UserService createUserService() {
-        return new HashMapUserService();
-    }
-
     protected Serialiser<T> createSerialiser() {
         return new NullSerialiser<>();
     }
 
     public Serialiser<T> getSerialiser() {
         return serialiser;
-    }
-
-    @Override
-    public ResourceService getResourceService() {
-        return resourceService;
-    }
-
-    @Override
-    public AuditService getAuditService() {
-        return auditService;
-    }
-
-    @Override
-    public PolicyService getPolicyService() {
-        return policyService;
-    }
-
-    @Override
-    public UserService getUserService() {
-        return userService;
-    }
-
-    @Override
-    public CacheService getCacheService() {
-        return cacheService;
-    }
-
-    @Override
-    public PalisadeService getPalisadeService() {
-        return palisadeService;
     }
 }
